+++ authlete/django/handler/spi/token_request_handler_spi_adapter.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2024 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -28,4 +28,12 @@ class TokenRequestHandlerSpiAdapter(TokenRequestHandlerSpi):
 
 
     def getProperties(self):
+        return None
+
+
+    def tokenExchange(self, tokenResponse):
+        return None
+
+
+    def jwtBearer(self, tokenResponse):
         return None