+++ authlete/django/handler/token_request_handler.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2024 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,7 +16,6 @@
 
 
 from authlete.django.handler.token_request_base_handler import TokenRequestBaseHandler
-from authlete.django.web.basic_credentials              import BasicCredentials
 from authlete.django.web.request_utility                import RequestUtility
 from authlete.django.web.response_utility               import ResponseUtility
 from authlete.dto.token_action                          import TokenAction
@@ -44,8 +43,8 @@ class TokenRequestHandler(TokenRequestBaseHandler):
     def handle(self, request):
         """Handle a token request.
 
-        This method calls Authlete's /api/auth/token API and conditionally
-        either /api/auth/token/issue API or /api/auth/token/fail API.
+        This method calls Authlete's /auth/token API and conditionally
+        either /auth/token/issue API or /auth/token/fail API.
 
         Args:
             request (django.http.HttpRequest)
@@ -57,15 +56,8 @@ class TokenRequestHandler(TokenRequestBaseHandler):
             authlete.api.AuthleteApiException
         """
 
-        # Request Body and Authorization Header
-        params = RequestUtility.extractRequestBody(request)
-        auth   = RequestUtility.extractAuthorization(request)
-
-        # BasicCredentials that represents the value of the Authorization header.
-        credentials = self.__parseCredentials(auth)
-
-        # Call Authlete's /api/auth/token API.
-        res = self.__callTokenApi(params, credentials)
+        # Call Authlete's /auth/token API.
+        res = self.__callTokenApi(request)
 
         # 'action' in the response denotes the next action which the
         # implementation of the token endpoint should take.
@@ -74,59 +66,75 @@ class TokenRequestHandler(TokenRequestBaseHandler):
         # The content of the response to the client application.
         content = res.responseContent
 
+        # Additional HTTP headers.
+        headers = self.__prepareHeaders(res)
+
         if action == TokenAction.INVALID_CLIENT:
             # 401 Unauthorized.
             return ResponseUtility.unauthorized(
-                'Basic realm="token"', content)
+                'Basic realm="token"', content, headers)
         elif action == TokenAction.INTERNAL_SERVER_ERROR:
             # 500 Internal Server Error
-            return ResponseUtility.internalServerError(content)
+            return ResponseUtility.internalServerError(content, headers)
         elif action == TokenAction.BAD_REQUEST:
             # 400 Bad Request
-            return ResponseUtility.badRequest(content)
+            return ResponseUtility.badRequest(content, headers)
         elif action == TokenAction.PASSWORD:
             # Process the token request whose flow is "Resource OWner
             # Password Credentials".
-            return self.__handlePassword(res)
+            return self.__handlePassword(res, headers)
         elif action == TokenAction.OK:
             # 200 OK
-            return ResponseUtility.okJson(content)
+            return ResponseUtility.okJson(content, headers)
+        elif action == TokenAction.TOKEN_EXCHANGE:
+            # Process the token exchange request (RFC 8693)
+            return self.__handleTokenExchange(res, headers)
+        elif action == TokenAction.JWT_BEARER:
+            # Process the token request which uses the grant type
+            # urn:ietf:params:oauth:grant-type:jwt-bearer (RFC 7523).
+            return self.__handleJwtBearer(res, headers)
+        elif action == TokenAction.ID_TOKEN_REISSUABLE:
+            # The flow of the token request is the refresh token flow
+            # and an ID token can be reissued.
+            return self.__handleIdTokenReissuable(res, headers)
         else:
             # 500 Internal Server Error
-            # /api/auth/revocation API returns an unknown action.
-            return self.unknownAction('/api/auth/revocation')
+            # /auth/token API returns an unknown action.
+            return self.unknownAction('/auth/token')
 
 
-    def __parseCredentials(self, auth):
-        if isinstance(auth, BasicCredentials):
-            return auth
-
-        if isinstance(auth, str):
-            return BasicCredentials.parse(auth)
-
-        # userId = None, password = None
-        return BasicCredentials(None, None)
-
-
-    def __callTokenApi(self, parameters, credentials):
-        if parameters is None:
-            # Authlete returns different error coes for None and an empty
-            # string. None is regarded as a caller's error. An empty string
-            # is regarded as a client application's error.
-            parameters = ''
-
-        # Create a request for /api/auth/token API.
+    def __callTokenApi(self, request):
         req = TokenRequest()
-        req.parameters   = parameters
+
+        # The request parameters.
+        req.parameters = RequestUtility.extractRequestBody(request) or ''
+
+        # The request may contain the basic authentication for client_secret_basic.
+        credentials      = RequestUtility.extractBasicCredentials(request)
         req.clientId     = credentials.userId
         req.clientSecret = credentials.password
-        req.properties   = self._spi.getProperties()
+
+        # The request may contain a client certificate.
+        req.clientCertificate = RequestUtility.extractClientCert(request)
+
+        # The request may contain a DPoP proof JWT.
+        req.dpop = request.headers.get('DPoP')
+
+        # Other parameters
+        req.properties = self._spi.getProperties()
 
         # Call /api/auth/token API.
         return self.api.token(req)
 
 
-    def __handlePassword(self, response):
+    def __prepareHeaders(self, res):
+        if res.dpopNonce is not None:
+            return { 'DPoP-Nonce': res.dpopNonce }
+
+        return None
+
+
+    def __handlePassword(self, response, headers):
         # The ticket to call Authelte's /api/auth/token/* API.
         ticket = response.ticket
 
@@ -141,7 +149,51 @@ class TokenRequestHandler(TokenRequestBaseHandler):
         if subject is None:
             # The credentials are invalid. Nothing is issued.
             return self.tokenFail(
-                ticket, TokenFailReason.INVALID_RESOURCE_OWNER_CREDENTIALS)
+                ticket, TokenFailReason.INVALID_RESOURCE_OWNER_CREDENTIALS, headers)
 
         # Issue tokens.
-        return self.tokenIssue(ticket, subject, self._spi.getProperties())
+        return self.tokenIssue(ticket, subject, self._spi.getProperties(), headers)
+
+
+    def __handleTokenExchange(self, tokenResponse, headers):
+        # Let the SPI implementation handle the token request.
+        response = self._spi.tokenExchange(tokenResponse)
+
+        # If the SPI implementation has prepared a token response, it is used.
+        # Otherwise, a token response with "error":"unsupported_grant_type" is
+        # returned.
+        return self.__useOrUnsupported(response)
+
+
+    def __handleJwtBearer(self, tokenResponse, headers):
+        # Let the SPI implementation handle the token request.
+        response = self._spi.jwtBearer(tokenResponse)
+
+        # If the SPI implementation has prepared a token response, it is used.
+        # Otherwise, a token response with "error":"unsupported_grant_type" is
+        # returned.
+        return self.__useOrUnsupported(response)
+
+
+    def __handleIdTokenReissuable(self, tokenResponse, headers):
+        # TODO: Support ID token reissuance.
+
+        # Note that calling ResponseUtility.ok() here will result in that the
+        # token endpoint behaves in the same way as before and no ID token is
+        # returned.
+        return ResponseUtility.ok(tokenResponse.responseContent, headers)
+
+
+    def __useOrUnsupported(self, response):
+        if response is not None:
+            return response
+
+        # Generate a token response that indicates that the grant type is not
+        # supported.
+        #
+        #     400 Bad Request
+        #     Content-Type: application/json
+        #
+        #     {"error":"unsupported_grant_type"}
+        #
+        return ResponseUtility.badRequest('{"error":"unsupported_grant_type"}')